Correct some more compilation warnings.  Added some additional Caret Assert Macros, CaretAssertDebugCompileOnly() and CaretAssertParameterDebugCompileOnly() that can be used to 'disable' variables and parameters that are only use in other CaretAssert() statements.

+++ src/Algorithms/AlgorithmNodesInsideBorder.cxx
@@ -763,9 +763,9 @@ AlgorithmNodesInsideBorder::findNodeFurthestFromConnectedPathCenterOfGravity(con
     
     if (sumCount >= 1.0) {
         const float cog[3] = {
-            sumX / sumCount,
-            sumY / sumCount,
-            sumZ / sumCount
+            (float)(sumX / sumCount),
+            (float)(sumY / sumCount),
+            (float)(sumZ / sumCount)
         };
         
         /*
